Move gammapy/spectrum/utils.py to modeling
