+++ gammapy/modeling/models/spectrum/__init__.py
@@ -3,3 +3,4 @@
 from .core import *
 from .cosmic_ray import *
 from .crab import *
+from .utils import *
