+++ gammapy/modeling/models/spectrum/crab.py
@@ -1,12 +1,7 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 import numpy as np
 from astropy import units as u
-from gammapy.modeling.models.spectrum.models import (
-    ExponentialCutoffPowerLaw,
-    LogParabola,
-    PowerLaw,
-    SpectralModel,
-)
+from .core import ExponentialCutoffPowerLaw, LogParabola, PowerLaw, SpectralModel
 
 __all__ = ["create_crab_spectral_model"]
