+++ gammapy/modeling/models/spectrum/utils.py
