+++ gammapy/spectrum/__init__.py
@@ -8,4 +8,3 @@ from .flux_point import *
 from .phase import *
 from .reflected import *
 from .sensitivity import *
-from .utils import *
